requirements.txt
python-pptx == 0.6.18
PyMuPDF == 1.17.0
comtypes == 1.1.7
PyPDF2 == 1.26.0